+++ api/app/helpers/spree/api/api_helpers.rb
@@ -24,6 +24,7 @@ module Spree
         :inventory_unit_attributes,
         :return_authorization_attributes,
         :creditcard_attributes,
+        :payment_source_attributes,
         :user_attributes,
         :property_attributes,
         :stock_location_attributes,
@@ -128,6 +129,10 @@ module Spree
       @@creditcard_attributes = [
         :id, :month, :year, :cc_type, :last_digits, :name,
         :gateway_customer_profile_id, :gateway_payment_profile_id
+      ]
+
+      @@payment_source_attributes = [
+        :id, :month, :year, :cc_type, :last_digits, :name
       ]
 
       @@user_attributes = [:id, :email, :created_at, :updated_at]