+++ api/CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Spree 2.3.0 (unreleased) ##
 
-* refactor the api to use a general importer in core gem.
+*   api/orders/show now display credit cards as source under payment
+
+    Washington Luiz
+
+*   refactor the api to use a general importer in core gem.
+
     Peter Berkenbosch